Moved draganimation
